requirements.txt
appdirs==1.4.3
diff-match-patch==20121119
Django==1.10.4
django-crispy-forms==1.6.1
django-debug-toolbar==1.7
django-filter==1.0.1
django-import-export==0.5.1
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
packaging==16.8
PyJWT==1.4.2
PyMySQL==0.7.9
pyparsing==2.2.0
pytz==2016.10
requests==2.13.0
six==1.10.0
sqlparse==0.2.3
tablib==0.11.4
